requirements.txt
discord.py
tldextract
sportsreference